feat(toolbar): add dashboard toolbar
